refs #refactor-rbac WIP refactor rbac into standalone classes
