+++ .gitignore
@@ -2,3 +2,4 @@
 /vendor
 /.idea
 composer.lock
+.php_cs.cache
