+++ tests/TestCase/Rbac/Rules/OwnerTest.php
@@ -8,7 +8,7 @@
  * @copyright Copyright 2010 - 2017, Cake Development Corporation (https://www.cakedc.com)
  * @license MIT License (http://www.opensource.org/licenses/mit-license.php)
  */
-namespace CakeDC\Auth\Auth\Rules;
+namespace CakeDC\Auth\Rbac\Rules;
 
 use Cake\Http\ServerRequest;
 use Cake\Network\Request;
